pyproject.toml
[tool.poetry]
name = "conway"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pygame-gui = "^0.6.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
